requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
environs==9.5.0
fastapi==0.104.1
fastapi_cors==0.0.6
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.1
h11==0.14.0
idna==3.6
imageio==2.33.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tifffile==2023.9.26
torch==2.1.1
torchaudio==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
